+++ CHANGELOG.md
@@ -2,7 +2,14 @@ Change Log
 ==========
 
 
-Version 1.0 *(2014-09-29)*
---------------------------
+Version 1.1.0 *(2014-10-06)*
+----------------------------
+
+* Convert `PrintView` from `TextView` to `ImageView`.
+* Convert `PrintButton` from `Button` to `ImageButton`.
+
+
+Version 1.0.0 *(2014-09-29)*
+----------------------------
 
 Initial release.